fix logic and add checking of current user while deleting

+++ Pospex/Pospex/Controllers/UsersController.cs
@@ -1,26 +1,46 @@
-﻿using Microsoft.AspNetCore.Identity;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Pospex.Models;
 using Pospex.ViewModels;
+using System;
+using System.Data;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Xml.Linq;
 
 namespace Pospex.Controllers
 {
     public class UsersController : Controller
     {
         UserManager<User> _userManager;
+        SignInManager<User> _signInManager;
 
-       
-        public UsersController(UserManager<User> userManager)
+
+        public UsersController(UserManager<User> userManager, SignInManager<User> signInManager)
         {
             _userManager = userManager;
+            _signInManager = signInManager;
         }
 
-        public IActionResult Index() => View(_userManager.Users.ToList());
+        public async Task<IActionResult> Index()
+        {
+            if (User.Identity.IsAuthenticated)
+            {
 
+                User currentUser = await _userManager.FindByEmailAsync(User.Identity.Name);
+                var userRoles = await _userManager.GetRolesAsync(currentUser);
+
+                ViewBag.UserRoles = userRoles;
+            }
+            return View(_userManager.Users.ToList());
+        }
+
+
+        [Authorize]
         public IActionResult Create() => View();
 
+        [Authorize(Roles = "admin")]
         [HttpPost]
         public async Task<IActionResult> Create(CreateUserViewModel model)
         {
@@ -43,6 +63,7 @@ namespace Pospex.Controllers
             return View(model);
         }
 
+        [Authorize(Roles = "admin")]
         public async Task<IActionResult> Edit(string id)
         {
             User user = await _userManager.FindByIdAsync(id);
@@ -54,6 +75,7 @@ namespace Pospex.Controllers
             return View(model);
         }
 
+        [Authorize(Roles = "admin")]
         [HttpPost]
         public async Task<IActionResult> Edit(EditUserViewModel model)
         {
@@ -82,15 +104,37 @@ namespace Pospex.Controllers
             return View(model);
         }
 
+        [Authorize(Roles = "admin")]
         [HttpPost]
         public async Task<ActionResult> Delete(string id)
         {
             User user = await _userManager.FindByIdAsync(id);
-            if (user != null)
+
+            User currentUser = await _userManager.FindByEmailAsync(User.Identity.Name);
+
+            if (user != null && currentUser != null)
             {
-                IdentityResult result = await _userManager.DeleteAsync(user);
+                if (currentUser != user)
+                {
+                    IdentityResult result = await _userManager.DeleteAsync(user);
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    var userRoles = await _userManager.GetRolesAsync(currentUser);
+                    ViewBag.UserRoles = userRoles;
+                    ViewBag.Message = "Error: Can not delete current user!";
+                    return View("Index", _userManager.Users.ToList());
+                }
+
             }
-            return RedirectToAction("Index");
+            else
+            {
+                var userRoles = await _userManager.GetRolesAsync(currentUser);
+                ViewBag.UserRoles = userRoles;
+                ViewBag.Message = "Error: User is not found!";
+                return View("Index", _userManager.Users.ToList());
+            }
         }
     }
 }
